Moved FieldSpecTests to correct package
